confdb: simplify matching using accessor type (#15654)
* confdb: simplify matching using accessor type

Use parsePathIntoAccessors to parse user supplied paths, like the
previous refactor did for view paths. This converts them into []accessor,
which we can then use when matching user paths with the assertion's
request paths. This simplifies the matching a bit because we can user
the helpers and the underlying key type to provide the same assurances
as before.
This also changes the validationOptions to more explicitly represent
that a path being validation and parsed can come from an assertion or a
user.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* confdb: test each sub-type key rejects wrong match

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>