cmd/snap: improve progress reporting when installing and refreshing snaps
When installing or refreshing snaps, and a snap pulls in additional
dependencies, the progress reporting gets stuck on 'Auto-connecting of
eligible slots and plugs'. The snap command is indeed waiting for
auto-connect task to complete since its status is still 'Doing', however
internally on the daemon side, the task is getting retried waiting for
prerequisites to finish installing/refreshing. As such, the actual
progress is somewhat invisible to the user and the installation refresh
command may appear stuck.

Attempt to resolve it by adding a simple heuristic to consider other
tasks in 'Doing' status if the currently considered task is auto-connect.

Related: SNAPDENG-35137

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>