overlord/notices: add notice manager and notice backend interface
Define a `NoticeManager` struct and a `NoticeBackend` interface.
Particular backends which implement the `NoticeBackend` interface may be
registered with a `NoticeManager` as providers of a particular notice
type, along with a unique namespace which must occur as a prefix to the
ID of any notice of that type produced by that backend.

A particular backend may be registered as a provider of more than one
notice type, as long as the namespace prefix registered for that backend
and type is unique. Similarly, multiple backends may be registered as
providers of the same notice type, as long as each has a unique
namespace prefix.

Snapd state is an implicit backend for any notice types which have no
registered backends. Notices from state have no namespace prefix, while
notices from all other backends must have a namespace prefix matching
that which was declared when the backend was registered for the
corresponding notice type.

The `NoticeManager` is responsible for deciding which backends are
capable of providing notices matching a given request, and only querying
those backends. The notice type in the filter is used to select relevant
backends for calls to `Notices` and `WaitNotices`, while the ID prefix
is used to select the relevant backend (if any) for calls to `Notice()`.

As a result, a request for notices of a type only provided by one
backend, for example, does not need to touch any other backend.
Importantly, this means that state lock is not required to complete
requests for notice types or IDs which are not provided by state.

The `NoticeManager` ensures that the existing invariants around notices
remain true. Namely, when the manager returns a list of notices from a
call to `Notices()` or `WaitNotices()`, it is guaranteed that there can
be no new notice recorded with a timestamp earlier than the
`lastRepeated` timestamp of the last notice in that list. (This did not
and still does not hold true for new notices which have a timestamp
explicitly specified.)

The `NoticeManager` provides a method to get a unique monotonically
increasing notice timestamp from the state (without holding state lock),
and `NoticeBackend`s must use it to get a timestamp whenever they create
or re-record a notice.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>