osutil/vfs: add tests for mount propagation behavior
We're interested in seeing what properties mount propagation settings
have on on new mount points that appear underneath. Create four new
tests that exercise this: one for private, shared, slave and unbindable
propagation.

A pair of awk scripts help us with the non-deterministic nature of the
peer group IDs and with printing a subset of mountinfo data.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>