i/prompting: add struct for holding current information
Define a new `prompting.At` struct which contains information about the
state at a current moment in time. For now, this just contains a `Time`
field which is populated with the timestamp of that moment in time.

In the future, however, this new struct can be extended to include other
relevant fields, such as the current user session ID.

This new struct is used in various places where a `currTime` timestamp
was previously passed, and also when validating the expiration of a
`RulePermissionEntry`.

Additionally, `IsRequestAllowed` was fixed to define a single current
timestamp (via the new `prompting.At` struct) which is then used when
checking all permissions. Previously, a new current timestamp was used
for each permission, which was a bug.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>