cmd/snap-confine: mount exported sources
Mount paths exported by nvidia userspace interfaces to the snap mountspace on
classic. This is done to keep compatibility with snaps using the opengl
interface, where we used to mount system libraries in the snap mountspace. Now,
however, these libraries could come from another snaps, so we need to use those
when found.