o/state: change notice Before filter to BeforeOrAt (#15700)
* o/state: change notice Before filter to BeforeOrAt

A `Before` notice filter was introduced in #15601 to allow querying for
notices before a given timestamp. The intent was (in the future) to
allow multiple notice backends to be queried in parallel, and then when
one returned a notice, re-query all backends for any notices with a
timestamp before the returned notice's last-repeated timestamp. But the
problem is that then that notice is itself omitted, since it does not
occur before its own timestamp.

To simplify things for this intended use-case, this commit changes the
`Before` filter into a `BeforeOrAt` filter, which matches notices with a
last-repeated timestamp which is before or equal to the `BeforeOrAt`
timestamp.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: reverse BeforeOrAt filter matching for clarity

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>