o/devicestate: use separate lanes for essential and non-essential snaps during seeding (#15672)
* o/devicestate, tests: use separate lanes for essential and non-essential snaps during seeding

This allows a non-essential snap to fail to be installed during seeding,
without undoing the installation of the essential snaps. On the next
ensure, we will retry seeding. If the non-essential snap installation
succeeds on any subsequent tries, then seeding will succeed.

* o/hookstate, o/snapstate: swap to a special case for seeding when organizing hook tasks

Since seeding uses lanes, we cannot just look for final tasks in the
same lane as the hook task. However, we don't want to unilaterally look
for all tasks in the change for final tasks. We keep this behavior as a
special case for seeding changes.