pipewire: Allow pipewire as system daemon (#15235)
Currently, pipewire creates the sockets at XDG_DESKTOP_RUNTIME;
this works fine for Core Desktop, where pipewire is launched as
an user daemon. Unfortunately, this doesn't work for Core,
because using users is discouraged. Also, even if the root user
were created (to have the /run/user/0/snap.pipewire folder), it
wouldn't work because the user daemon socket would be already
there, so the system daemon socket couldn't be created in that
folder.

The solution is to use XDG_DESKTOP_RUNTIME for the user daemon
socket, and SNAP_COMMON for the system daemon (it must be a
folder that supports socket activation), and then let the
developer choose which daemon to use by setting PULSE_SERVER
and PIPEWIRE_RUNTIME_DIR environment variables.

Unfortunately, this still has a problem: other snaps can't get
access to that folder, so they wouldn't be able to connect to
the daemons, no matter that they have connected the corresponding
plug.

This patch fixes this, by giving access to
/var/snap/SLOT_SNAP_NAME/common/pulse[/native] and
/var/snap/SLOT_SNAP_NAME/common/pipewire-[0-9] when the
audio-playback and pipewire interfaces are connected.