o/notices: remove map from backend and type to namespace
The intention of this map was to prevent a given backend/type pair from
being registered multiple times with multiple namespaces. However,
there's no particular reason why this shouldn't be allowed, given that a
backend can register multiple notice types with the same namespace.

An original motivation for this map was also to allow the notice manager
to validate that a given notice of a particular type from a particular
backend had an ID prefix matching the registered namespace. However, the
`validateNotice` closure returned after registration does this, and it
is not the role of the notice manager to validate notices, but instead
the role of each notice backend to ensure that it is producing
well-formed notices.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>