intefaces/builtin, cmd/snap: Kerberos tickets support (#15519)
* cmd/snap: Set Kerberos cache to hostfs path.

SSSD creates the user's Kerberos ticket in /tmp and sets
the KRB5CCNAME to point to that file.

Snaps have a private /tmp so they can't reach that file.

So prepend /var/lib/snapd/hostfs/ to KRB5CCNAME when running the snap,
which will still require a interface (kerberos-tickets) to read the
tickets from that path.

* interfaces: add kerberos-tickets interface

SSSD creates the user's Kerberos ticket in /tmp.

The previous commit ensures KRB5CCNAME is wired through the
hostfs' tmp directory so that the ticket can be located by the snap.

Still, Apparmor permissions to read and lock the ticket are needed,
which is what this interface provides.

* tests: Add interfaces-kerberos-tickets spread test.

* tests: give kerberos-tickets interface 08/2025 deadline.