osutil/vfs: introduce attachMount and detachMount
The two new helpers organize the code that is required for correct linkage
of mounts into a chain of parent-to-children and child-to-siblings as
well as child-to-parent.

The code for Mount, BindMount and Unmount are updated to use the new
helpers.

A small extra helper for allocating a MountID is also introduced, mainly
to hold the necessary explanation.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>