cmd/snap-confine: switch to root:root to operate v1 freezer
Snap-confine contains cgroup-v1-specific behavior: the use of the freezer group
for both application tracking and consistency guarantee for snap-update-ns that
may be concurrently modify the mount namespace.

The switch to capability-based snap-confine missed the fact that moving
processes within a cgroup is governed by additional constraints that no
capabilities can remedy. The process making the move must either be a global
root (not in the process namespace) or the UID of the process being moved (here
snap-confine moves itself) must match the UID of the target cgroup hierarchy.
This is not true in LXD containers that use user namespaces.

The failure was introduced in snapd 2.70 but due to a faulty test, was not
detected until after the release.

The fix alters the freezer join logic to raise UID/GID to root, to create the
cgroup directory (GID) and move the process (UID) and then revert the changes.
This approach was selected by eliminating alternatives driven by the ultimate
constraint that the kernel imposes.

```c
/*
	 * Even if we're attaching all tasks in the thread group, we only need
	 * to check permissions on one of them. Check permissions using the
	 * credentials from file open to protect against inherited fd attacks.
	 */
	cred = of->file->f_cred;
	tcred = get_task_cred(task);
	if (!uid_eq(cred->euid, GLOBAL_ROOT_UID) &&
	    !uid_eq(cred->euid, tcred->uid) &&
	    !uid_eq(cred->euid, tcred->suid))
		ret = -EACCES;
```

Any of the three alternatives grants us access. We cannot rely on global root
as snapd is required to work in user namespaces.  We must then attempt
snap-confine to match itself (second comparison) while also having enough
permissions to create the freezer sub-directory and open the cgroup.procs file.

In order to alter uid/gid we need the capabilities CAP_SETUID and CAP_SETGID.
Sadly, this brings us closer to the original permission set before the
capability switch in snap-confine. As a small remedy this code can be
eventually dropped when support for cgroup-v1 systems is removed.  To further
harden snap-confine, we are dropping the new capabilities from all the sets
early on startup, as soon as it is clear we are using cgroup-v2 mode.

Lastly fix the critical test that tested stable, released copy of snapd, making this
bug slip past CI/CD during the release process of snapd 2.70.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>