confdb: support lists in various checks and helpers
Export accessor and associated methods and helpers so it can be used in
exported methods. Change the storage path to be computed into []Accessor
early and use that in the databag and schema methods Get/Set/Unset and
SchemaAt. Since SchemaAt now has more information in the path, we can
use that to strengthen the schema/path checks there.
Adds list/indexed path support to the methods we use for early checks
and various things (Read|WriteAffectsEphemeral, GetViewsAffectedByPath,
etc).

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>