o/notices: add method to drain notices of a given type from state
When a new notice backend is registered, the intention may be for it to
be the sole provider of notices of a given type, and state should no
longer produce or manage those notices. As such, notices of that type
should be drained from state and re-added into that backend instead.

Since each notice backend has its own method to add new notices, which
may or may not be exported, the manager can't automatically drain and
import the notices into the backend as part of registration. As such,
this commit adds a new exported method to the manager which should be
called by the same caller responsible for registering a new notice
backend, if desired.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>