cmd: fix installation of snap-confine profile with 'make hack'
Fix how snap-confine profile is installed when using 'make hack'.
Install the appropriate profile for the target system. This avoids
situation when on Arch or openSUSE, one could end up with 2 AppArmor
profiles for snap-confine, one in usr.lib.snapd.snap-confine and another
in usr.lib.snapd.snap-confine.real. This causes a situation in which a
parallelized profile loading at boot time which would
nondeterministically put one or the other profile in effect.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>