image: improve handling of system-user extra assertions (#15671)
* image: improve handling of system-user extra assertions

Since e0e4f9b, it is possible to provide extra assertions to the image
being built. However, system-user assertions will not result in users
being created. This commit improves this special case, and restricts
system-user assertions provided this way to public key authentication
only.

Signed-off-by: Nicolas Harnois <nicolas.harnois@canonical.com>

* image: allow seeding system-user assertions only for dangerous models

If an extra system-user assertion is supplied at image build time, this
commit with reject it if the model grade is different from dangerous.

Signed-off-by: Nicolas Harnois <nicolas.harnois@canonical.com>

* image: fixes after review

Signed-off-by: Nicolas Harnois <nicolas.harnois@canonical.com>

---------

Signed-off-by: Nicolas Harnois <nicolas.harnois@canonical.com>