multiple: refactor lookup of integrity data
add a new NewIntegrityDataParamsFromRevision which takes a snap
revision as input and will return integrity data if found or an error
in the case where integrity data is not found or is of unsupported type.

The choice to ignore error if verity data is not found for a snap
remains with the caller. Ignoring this error in seed20 means that snaps
that don't have integrity data in their snap-revision assertions will
work.