osutil/vfs: replace [mount.mountPoint] with [mount.attachedAt]
The change is perhaps tiny but helps greatly in upcoming operations.
The main benefit is that the suffix of the mount is no longer interested
as to where it is exactly attached. Recursive bind, pivot root and mount
move operation all change that location. With the new design most of the
operations will no longer need to touch the new [mount.attachedAt] value,
but merely invalidate the mount point cache.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>