osutil/vfs: add support for mount propagation (WIP)
This applies small refactoring to the VFS package, to keep track of
propagation settings and to keep track of parent-child relationship
with a pointer on top of the existing ID.

The key new state is a pair of identifiers added to each mount: shared
and master. A non-zero value indicates that the given mount is a peer
(shared) or slave (master) to the group.

A lesser flag is also added: unbindable. Any mount can be unbindable,
private, shared, shared and slave, slave or unbindable.

Known TODOs:
 - Unbindable is not implemented fully
 - Unmounts do not propagate
 - We may need re-propagation from shared slave

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-34951
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>