daemon,overlord: use notice manager and query from daemon (#15704)
* daemon,overlord: use notice manager and query from daemon

Initialize a new `NoticeManager` after state is loaded from disk, and
expose it under a new `NoticeManager()` method on the overlord.

Then, for the notices endpoints which query existing notices, switch
from querying notices directly from state, and instead query via the
overlord's notice manager.

The goal of this is that if there is an API request for notices which
is satisfied by notice backends outside of state, then there is no need
to acquire state lock in order to process that notices request, and the
notice manager can instead only query the relevant notice backends.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* daemon,overlord: clarify comment around notice manager and state lock

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>