osutil/vfs: add MetaDataFS, WithSource for tests
WithSource provides the source field when looking at the textual
representation of the mount table, making test output easier to parse
for humans.

Implementation extends a similar helper file system wrapper so that now
we have a extensible meta-data source for any tests that want to
annotate file systems with readable information.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>