tests: fix detection of broken packagekit
Some versions of packagekit are affected by a bug that only manifests
itself when a proxy is configured (or even lack of proxy is requested).

Adjust the version-based check to look for the faulty behavior and
reduce the test to a subset that should pass on all the systems - an
offline backend check.

The bug in question is fixed by upstream packagkit commit:

commit 4626cc3bec7affb632122b1d3d4898d7a5bd0e7d
Author: Philip Withnall <pwithnall@endlessos.org>
Date:   Fri Oct 21 15:02:59 2022 +0100

    pk-engine: Fix checking D-Bus caller details when setting proxy

    This was broken in commit c8b531.

    This should fix errors like the following when calling `SetProxy()`:
    ```
    PackageKit-WARNING **: 14:50:25.673: failed to set proxy: GDBus.Error:org.gtk.GDBus.UnmappedGError.Quark._pk_2dengine_2derror_2dquark.Code3: setting the proxy failed: failed to get the uid
    ```

    Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>