o/notices: optionally drain notices on backend registration
When a new notice backend is registered, the intention may be for it to
be the sole provider of notices of a given type, and state should no
longer produce or manage those notices. As such, notices of that type
should be drained from state and re-added into that backend instead.

Add a new boolean parameter to `RegisterBackend` which, if true, will
drain all notices of the type newly registered to the backend, and
return them. The intention is that the caller will then migrate these
existing notices into that backend.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>