osutil/vfs/tests: expand bind-semantics measurements
There are two changes in one:

1) The test constructs A and B instead of a and b, then binds A/a to B/b
2) The test observers A, B and B/b

This makes it easier to see what B is and to see what exactly is going
on at A/a vs B/b. Bind-mounting a sub-directory no longer hides the
original mount, making the file system structure easier to grok (at
least for me).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>