o/state: support reading notices without holding state lock
Add a `RWMutex` to guard the state notices map, and no longer require
the state lock to be held just to read notices.

This means that callers of `state.Notices`, `state.Notice`, and
`state.WaitNotices` no longer need to hold the state lock.

As a result, the `state.noticeCond` locker is switched from locking
`state` itself to instead locking the `RWMutex`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>