o/state: avoid recursive read locking to prevent deadlock
Ensure that the notices `RWMutex` is never recursively locked for
reading. Since any call to `Lock` synchronizes before any subsequent
call to `RLock`, if a function (namely `WaitNotices`) attempts to call
`RLock` while it already holds the lock for reading, and another thread
has called `Lock`, then that new `RLock` call will block indefinitely,
preventing the function from releasing the original read lock, and
leading to deadlock.

Instead of calling `Notices()` directly from within `WaitNotices()`,
instead split out the body of `Notices` into `doNotices`, which does not
acquire the lock again. Since `flattenNotices` needs to acquire the lock
but does not need to sort the returned notices, split out the body of
`flattenNotices` into `filterNotices`, which is called by `doNotices`
(which sorts but does not lock) and `flattenNotices` (which locks but
does not sort).

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>