tests: run tests/nested/manual/hybrid-fde-recovery-keys for 25.10 again (#15840)
It seems that tests/nested/manual/hybrid-fde-recovery-keys stopped working
silently some time ago, which explains why it didn't fail when the 25.10
gadget seed disk size was increased.

It started not working in https://github.com/canonical/snapd/pull/15711
because the 2510 variant was removed and switched as default env vars
and it didn't run because there were no other variants that don't
override the relevant env vars.

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>