store: skip icon download if proxy is in use
If a store proxy is being used during snap installation, then skip
downloading the snap icon. As downloading an icon involves simply
getting a file from a fileserver (typically dashboard.snapcraft.io), the
store proxy is not designed to mediate this download. However, when a
proxy is in use, the expectation is that all snapd traffic should go
through the proxy. Thus, the icon download should be skipped.

Prior to this change, in airgapped environments using the store proxy,
it was possible that the icon download would hang for 15 minutes before
failing, thus delaying all snap install/refresh tasks dramatically. This
commit prevents this issue from occurring.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>