o/i/apparmorprompting: preserve expired notices on error
If an error occurs during `addNotice`, ensure that an existing expired
notice is not deleted from the ID map.

This isn't strictly necessary, but it's a matter of consistency, since
`BackendNotice(id)` doesn't currently check whether the notice it looks
up has expired. We want a failed `addNotice` call to have no side
effects, so the result of `BackendNotice(id)` should be the same before
and after the failed call.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>