tests: small improvements to mounts persist refresh content (#14722)
* tests: break instead of exiting

The test should have broken out of the loop, so that the "exit" message
is observed by the test regardless of the success or failure.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: redirect input to MATCH and NOMATCH

I think MATCH/NOMATCH does not cope well with arguments passsed directly,
like what one would to do grep.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: forcibly remove the keep-running flag

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: stat before/after

Show stat of /usr/share/fonts/foo-font of before / after the refresh of
the content snap. This is to ensure that after the test is over we can
still see the directory that we were expecting.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: more precise checks for missing paths

When the test fails, we'd like to know which path was actually missing.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: add debug section to mounts-persist-refresh-content-snap

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

---------

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>