confdb: refactor databag error usage
The PathError no longer makes sense since the only time we handle it
explicitly in confdb.go is when there's no data to read and there's
already a more specific error for that. The other cases are just generic
errors during traversal which we don't handle and don't need a type.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>