o/i/apparmorprompting: avoid alloc/sort in doNotices
For consistency regardless of whether a `UserID` field is given in the
notice filter, reuse the existing slice if notices are only from a
single user.

Similarly, don't sort the resulting notices if they are only from one
user.

If there are notices from two or more users, a new slice must be
allocated so that it can be later sorted without mutating the sorted
user notices slice for any user.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>