o/state,o/snapstate: make warnings backed by notices
Instead of separate storage and similar logic for managing warnings and
notices, reimplement warnings in terms of notices while preserving the
existing warnings API and behavior.

When adding a warning, record a notice of type `WarningNotice`. Warnings
and notices both have the concept of "repeat-after", but in the case of
Warnings, it really means show again, after the warning having been
shown and okayed. Accordingly, warnings also have a "last-shown"
timestamp, which is the time when the warning was last shown and okayed.

These two warnings-specific fields must be stored as part of the notice.
For this initial implementation, they are formatted as strings and
stored in the general-purpose "last-data" map, so as to avoid adding any
new dedicated fields to notices which are only of use to warning
notices. However, this approach comes with the consequence of requiring
manual round-tripping between time/duration and string. For simplicity,
it may be better to add a warning metadata sub-struct to the notice
struct, which can be omitted for non-warning notices.

Lastly, refresh inhibit warnings feature mutable messages, as the list
of inhibited snaps changes. Previously, the refresh inhibited warning
needed to be looked up by suffix, removed, and a new warning added with
the new message. Notices cannot be manually removed, so a new notice
needs to be recorded with the same key as the existing notice so as to
overwrite it. Thankfully, notice "last-data" is the perfect mechanism,
as it allows us to store a custom "details" value which we can use in
place of the notice key for warning messages, if present. For refresh
inhibit warnings, we can use a fixed key and display the "details" as
the warning message.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>