o/state,o/snapstate: reuse existing refresh inhibit warning notice
Since we're using notice data to store the warning "details" message
which can be updated when a warning is re-recorded, we don't need to
delete the existing warning before re-adding it.

When there are no more inhibited refreshes, however, we do still
"remove" the existing warning notice by re-recording it with an
"expire-after" value of 1 nanosecond. However, to guarantee that it is
already expired before the function to remove it has finished (assuming
the rest of the function could finish instantaneously), the timestamp
needs to be set to at least 1ns before the current time minus the
"expire-after" value, so 2ns before now.

Additionally, by setting the notice timestamp manually, we avoid bumping
the state's `lastNoticeTimestamp` as part of our attempt to effectively
remove the existing notice.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>