o/state: add question about reusing expired notice
Currently, if a notice is expired and we attempt to record a notice with
the same user/type/key, then that existing notice is re-recorded despite
the fact that it has already expired.

As a result, whether expired notices happen to have been pruned between
when the notice expired and when it is being recorded again has an
effect on whether it is reused or a new notice is created. This seems
inconsistent, so it may be preferable to always create a new notice if
the existing notice has expired.

However, when this change is made, `TestOldRepeatedWarning` fails, so I
think there are some unintended side effects of reused notices on
warnings at the moment, which should be investigated further.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>