i/b/xilinx-dma: support Multi Queue DMA (QDMA) IP
Support both qdma-pf and qdma-vf modes

When a queue is created and started, the driver will create a device
entry (such as `/dev/qdma01000-MM-0`) dynamically for applications
to use.
The configuration entries will be in the PCIe device entries of
sysfs. For instance:
`/sys/devices/pci0000:00/0000:00:01.1/0000:01:00.0/qdma/`.

The hardware-observe interface may also be required for an
application to gather hardware information before accessing the
actual QDMA device.

Signed-off-by: Robert Liu <robert.liu@canonical.com>