s/a/notify: respond with full permission set user allowed
Previously, a snapd allow response to a request from the kernel only
ever allowed permissions which AppArmor initially allowed and those
which were initially denied and the user then explicitly allowed.
Notably, if the user allowed a broader set of permissions, those which
were not originally requested would be omitted from the response to the
kernel.

This commit changes this behavior. Now, include all permissions the user
explicitly allowed in the allowed permission mask sent back to the
kernel.

The rationale is that newer kernels support response caching at the file
descriptor level, so if the user replies to a request prompt by allowing
with lifespan "single", the response can be cached on the FD and future
syscalls on the same file descriptor will be handled by the cache
without resulting in another request prompt to snapd.

However, permissions are exposed to users in an abstracted form, e.g.
"read", "write", "execute" for the home interface. Thus, the "write"
permission actually includes many separate operations, including
create, write, lock, chmod, etc. If an application attempts to create a
file, write to it, and then chmod it, the user will get three prompts to
"write" the same file, assuming they reply with allow "single" each
time.

By replying with the full permission mask associated with the abstract
permission which the user allowed, we can leverage the file descriptor-
level cache and handle this sequence of "write" syscalls using a single
request prompt, even if the user replies with allow "single".

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>