many: read notices without state lock (#15819)
* o/state: support reading notices without holding state lock

Add a `RWMutex` to guard the state notices map, and no longer require
the state lock to be held just to read notices.

This means that callers of `state.Notices`, `state.Notice`, and
`state.WaitNotices` no longer need to hold the state lock.

As a result, the `state.noticeCond` locker is switched from locking
`state` itself to instead locking the `RWMutex`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/notices: avoid state lock when querying the state for notices

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* daemon: avoid state lock when querying the state for notices in test

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: use noticesMu.RLocker as the noticeCond locker

When `noticeCond.Wait` is called, it will attempt to unlock
`noticeCond.L`. Therefore, `noticeCond.L` must be the same lock which is
held during the call to `noticeCond.Wait`, which is `noticeMu.RLock`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: ensure noticeCond always uses noticesMu.RLocker

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: avoid recursive read locking to prevent deadlock

Ensure that the notices `RWMutex` is never recursively locked for
reading. Since any call to `Lock` synchronizes before any subsequent
call to `RLock`, if a function (namely `WaitNotices`) attempts to call
`RLock` while it already holds the lock for reading, and another thread
has called `Lock`, then that new `RLock` call will block indefinitely,
preventing the function from releasing the original read lock, and
leading to deadlock.

Instead of calling `Notices()` directly from within `WaitNotices()`,
instead split out the body of `Notices` into `doNotices`, which does not
acquire the lock again. Since `flattenNotices` needs to acquire the lock
but does not need to sort the returned notices, split out the body of
`flattenNotices` into `filterNotices`, which is called by `doNotices`
(which sorts but does not lock) and `flattenNotices` (which locks but
does not sort).

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: add test against notice wait deadlock

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: add comment explaining locks held during WaitNotices

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/state: adjust comment about notices RWMutex in WaitNotices

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>