spread, tests: split up fips snap and deb test suites
Spread does not properly hadle variants at the suite level, i.e. the
suites executing on a single host are not grouped by variant and then
their prepare steps aren't executed for each variant independently. This
causes the tests to run with unpredictable state, which depends entirely
on which variant happend to run first.

As a result, the FIPS test suite needs to be split up into 2 independent
suites, one for snapd deb and one for snapd snap.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>