spread: run FIPS test suite with actual DEB package (#15923)
* spread: run FIPS test suite with actual DEB package

Make sure that we're running FIPS test suite with the actual deb
package when needed.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* spread: do not skip the FIPS suite

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* tests/main/fips: do not skip the task when not using test keys

Test keys are irrelevant for the test. What matters is snapd deb being
rebuilt from source tree which is set up at the suite level.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* spread, tests: split up fips snap and deb test suites

Spread does not properly hadle variants at the suite level, i.e. the
suites executing on a single host are not grouped by variant and then
their prepare steps aren't executed for each variant independently. This
causes the tests to run with unpredictable state, which depends entirely
on which variant happend to run first.

As a result, the FIPS test suite needs to be split up into 2 independent
suites, one for snapd deb and one for snapd snap.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* github, tests: mappings and execution for fips and fips-deb suites

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* spread, tests/main/fips: use FIPS_ARTIFACT to indicate which artifact is relevant for the test

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* github: split fips spread tests into 2 separate groups ubuntu-fips-snap and ubuntu-fips-deb

The snapd deb packages are built in project prepare, so when 2 suites
expecting packages built differently, run on the same host, the result
is unpredictable and depends on which suite happens to run first.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* many: rename tests/fips -> tests/fips-snap

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* spread: consistent quoting

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* github/workflows/spread-tests: better selector for fips test groups

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>