store: use http(s) proxy for icon downloads, improve unit tests (#15948)
* store: use common client setup such that configured http(s) proxy is used if configured

Tiny refactor of the icon download code to use the common client setup
mechanism. This ensures that should there be a http(s) proxy configured
in the system and used by the client, the icon download code will use it
as well.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* store: add tests verifying that http(s) proxy configured for the store is used

Improve unit tests to mke sure that whenever a http(s) proxy was passed
in store configuration it will be used by actions and download code.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>