interfaces: add multipath-control interface
This interface allows a snap to communicate with multipath on the host
to manage device-mapper devices on the host.

It will be used both by `cinder-volume` and `openstack-hypervisor` snap
to allow mounting redoundant devices from enterprise arrays, such as
ISCSI target at first, and then later expended to NVMEoF and Fiber
Channel.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>