interfaces/builtin: replace failed checkbox-support interface
The checkbox-support interface was meant to be steam-like, in the sense that
the sandbox was largely disabled. What we missed during the original design was
that checkbox often runs snap applications, while steam does not. As such,
checkbox really needs a full sandbox escape, instead of a "sandbox light" mode
like we thought before.

The interface remains super-privileged but the nature of the implementation
changed entirely as the former implementation was never productised by
checkbox.

During the last couple of weeks we've designed and debugged another approach,
where checkbox uses systemd D-Bus APIs to break out of the sandbox. This was
both prototyped and developed into a small tool, plz-run -
https://gitlab.com/zygoon/plz-run - along with the checkbox developer
Massimiliano Girardi.

The new set of permissions allow checkbox to StartTransientUnit, KillUnit and
to listen to JobRemoved and PropertiesChanged signals.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-35415

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>