tests: post testing metrics
This is the first approach used to submit testing metrics for spread
executions.

For this purpose, we measure an empty test to calculate the pure
preparation and restoring time of a project, suite, or task.

The focus is mainly on the repeated parts of the code, where any change
could impact the overall testing time.