s/a/notify: protocol v5 support check only kernel tags support
Previously, the check for notify protocol v5 support checked that both
the kernel and the AppArmor parser supported metadata tags. However,
only kernel support for tags is required for notify protocol v5 to be
used. The parser is not relevant to the protocol used to communicate
directly between snapd and the kernel.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>