interfaces: add nvme-control interface
This interface allows a snap to manage NVMe controllers and namespaces
through the kernel's NVMe subsystem on the host, including NVMe over
Fabrics (NVMe-oF) connections.

It will be used by storage management snaps like `cinder-volume` and
`openstack-hypervisor` to configure and monitor NVMe storage arrays,
NVMe-oF targets, and software-defined storage solutions in enterprise
environments.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>